+++ draft/test_sur.py
@@ -3,8 +3,7 @@ Created: Mon Jul  8 12:12:59 2019
 @author: Christopher Albert <albert@alumni.tugraz.at>
 """
 
-from profit.sur import Surrogate
-from profit.sur.backend.gp import GPSurrogate
+from profit.sur.gp import GPSurrogate
 
 a = GPSurrogate()
