+++ examples/algae/test_algae.py
@@ -10,7 +10,7 @@ import os
 from collections import OrderedDict
 from chaospy import J, generate_quadrature, orth_ttr, fit_quadrature, E, Std, descriptives, Normal
 
-from profit.sur.backend.gp import GPFlowSurrogate
+from profit.sur.gp import GPFlowSurrogate
 
 import pickle
